refactor table to be at same level as plots as base infra
